src/test/java/com/github/malapert/wkt/metadata/ScopeTest.java
/* 
 * Copyright (C) 2016-2019 Jean-Christophe Malapert
 *
 * JWkt is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3.0 of the License, or (at your option) any later version.
 *
 * JWkt is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public
 * License along with this library; if not, write to the Free Software
 * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
 * MA 02110-1301  USA 
 */
package com.github.malapert.wkt.metadata;

import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author malapert
 */
public class ScopeTest {
    
    public ScopeTest() {
    }
    
    @BeforeClass
    public static void setUpClass() {
    }
    
    @AfterClass
    public static void tearDownClass() {
    }
    
    @Before
    public void setUp() {
    }
    
    @After
    public void tearDown() {
    }

    /**
     * Test of getDescription method, of class Scope.
     */
    @Test
    public void testGetDescription() {
        Scope instance = new Scope("Large scale topographic mapping and cadastre.");
        String expResult = "Large scale topographic mapping and cadastre.";
        String result = instance.getDescription();
        assertEquals("getDescription", expResult, result);
    }

    /**
     * Test of setDescription method, of class Scope.
     */
    @Test
    public void testSetDescription() {
        String description = "test_description";
        Scope instance = new Scope("test");
        instance.setDescription(description);
        assertEquals("setDescription", description, instance.getDescription());
    }
    
}
